moved calendar to the popup file
