--- src/components/User/UserCalendar/UserCalendar.js
@@ -1,22 +1,0 @@
-import React, { forwardRef, useState } from 'react';
-import DatePicker from 'react-datepicker';
-
-import 'react-datepicker/dist/react-datepicker.css';
-import css from './UserCalendar.module.css';
-
-export const CalendarDropdown = () => {
-  const [startDate, setStartDate] = useState(new Date());
-  const CustomInput = forwardRef(({ value, onClick }, ref) => (
-    <button className={css.customInput} onClick={onClick} ref={ref}>
-      {value}
-    </button>
-  ));
-
-  return (
-    <DatePicker
-      selected={startDate}
-      onChange={date => setStartDate(date)}
-      customInput={<CustomInput />}
-    />
-  );
-};
