--- src/components/User/UserCalendar/UserCalendar.module.css
@@ -1,13 +1,0 @@
-.customInput {
-  background-color: var(--main-color-blue);
-  border: none;
-  border-radius: 8px;
-  padding: 6px 12px;
-  cursor: pointer;
-  color: #fff;
-  text-align: center;
-  font-family: Inter, sans-serif;
-  font-size: 14px;
-  font-weight: 700;
-  text-transform: uppercase;
-}
